Check to see if current user can see all settings
We should check access with current_user_can, the way we do it today can create problems for users being superadmins for example and WP discourages checking against roles, rather than checking for capabilities.

We had an example on a multisite where the superadmin got false for $allow_edition and thus couldn't access the plugin settings fully.

manage_otions will render it for admins and super admins (superadmins always get true on current_user_can)
If we want editors to have access as well, edit_pages would probably be a good capability. If we check against edit_pages we need to supply the function with a $object_id param.